Separate header to more smaller parts
