+++ src/components/headerSearch.tsx
@@ -1,18 +1,10 @@
 import Image from "next/image";
-import { useState } from "react";
 import { Search } from "react-feather";
+import { getMovieByLimitTypeGenre } from "@/services/movieService";
 
-const placeholderMovies = [
-  { name: "Movie 1", poster_path: "jf3YO8hOqGHCupsREf5qymYq1n.jpg", id: 1, year: "2000" },
-  { name: "Movie 2", poster_path: "jf3YO8hOqGHCupsREf5qymYq1n.jpg", id: 2, year: "2000" },
-  { name: "Movie 3", poster_path: "jf3YO8hOqGHCupsREf5qymYq1n.jpg", id: 3, year: "2000" },
-];
-
-export const HeaderSearch = () => {
-  const placeholderResults = "";
-  const [search, setSearch] = useState<string>("");
-
-  console.log(placeholderResults);
+export const HeaderSearch = async () => {
+  const placeholderResults = await getMovieByLimitTypeGenre(4, "popular", undefined);
+  const search = "";
 
   return (
     <div className="header-searchbar">
@@ -21,7 +13,6 @@ export const HeaderSearch = () => {
         data-cy="search-input"
         className="searchbar-input"
         placeholder="Search..."
-        onChange={e => setSearch(e.target.value)}
       />
       <button className="button-transparent">
         <Search
@@ -30,26 +21,26 @@ export const HeaderSearch = () => {
         />
       </button>
 
-      {search !== "" && (
-        <div className="searchbar-results">
-          {placeholderMovies.map(movie => (
-            <button
-              key={movie.id}
-              className="searchbar-movie-result"
-            >
-              <Image
-                alt={movie.name}
-                src={`https://image.tmdb.org/t/p/w500/${movie.poster_path}`}
-                width={35}
-                height={50}
-              />{" "}
-              {movie.name} ({movie.year})
-            </button>
-          ))}
+      <div className="searchbar-results">
+        {placeholderResults.map(movie => (
+          <button
+            key={movie.id}
+            className="searchbar-movie-result"
+          >
+            <Image
+              alt={movie.title}
+              src={`https://image.tmdb.org/t/p/w500/${movie.poster_path}`}
+              width={35}
+              height={50}
+            />
+            <div className="result-title">
+              <span>{movie.title}</span> ({movie.release_date?.getFullYear()})
+            </div>
+          </button>
+        ))}
 
-          <button className="searchbar-page-button">See all results for &quot;{search}&quot;</button>
-        </div>
-      )}
+        <button className="searchbar-page-button">See all results for &quot;{search}&quot;</button>
+      </div>
     </div>
   );
 };
